Rename BoardPlayerGroup to ViewerGroup & move to state
