+++ scoreboard/spigot/build.gradle.kts
@@ -9,6 +9,7 @@ dependencies {
     api(project(":scoreboard"))
     api(project(":common"))
     api(project(":spigot"))
+    api(project(":state:state-spigot"))
 
     compileOnly(libs.paper)
